move utils to internal package
