+++ nix/src/main/java/org/gnode/nix/File.java
@@ -4,6 +4,7 @@ import org.bytedeco.javacpp.IntPointer;
 import org.bytedeco.javacpp.Loader;
 import org.bytedeco.javacpp.Pointer;
 import org.bytedeco.javacpp.annotation.*;
+import org.gnode.nix.internal.Utils;
 
 import java.util.ArrayList;
