+++ nix/src/main/java/org/gnode/nix/internal/Utils.java
@@ -1,4 +1,4 @@
-package org.gnode.nix;
+package org.gnode.nix.internal;
 
 import org.bytedeco.javacpp.DoublePointer;
 import org.bytedeco.javacpp.IntPointer;
